+++ clients/pipeline-service/pms-sdk-core/src/main/java/io/harness/waiter/AsyncWaitEngineImpl.java
@@ -28,7 +28,7 @@ public class AsyncWaitEngineImpl implements AsyncWaitEngine {
 
   @Override
   public void waitForAllOn(
-      NotifyCallback notifyCallback, ProgressCallback progressCallback, List<String> correlationIds, int timeout) {
+      NotifyCallback notifyCallback, ProgressCallback progressCallback, List<String> correlationIds, long timeout) {
     waitNotifyEngine.waitForAllOn(
         publisherName, notifyCallback, progressCallback, correlationIds, Duration.ofMillis(timeout));
   }
